Add a python 3.7 image.
I commented out the list of tags and added one for 3.7. The patch versions we have specified are no longer
available on docker hub (only the latest patch version for a Major.Minor build is available) and I didn't want
to update our images without testing.
For available images see:
https://hub.docker.com/_/python